feat: unify payslips into one page (admin)
