+++ src/components/payslips/payslip.ts
