+++ src/constants.ts
@@ -5,5 +5,4 @@ export const LINKS = {
   timelogshistory: "/time-logs-history",
   registeruser: "/register",
   payslips: "/payslips",
-  viewpayslips: "/view-payslips",
 };
